+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 ### Changed
+- Replacing `EntitySource.safeMapGet` by `EntitySource.getOrThrow` [#828](https://github.com/ie3-institute/PowerSystemDataModel/issues/828)
 
 ## [4.1.0] - 2023-11-02